client: first pass at automatic installation
Image operations will now incur an automatic installation attempt. Such
will fail on anything other than single-node clusters. This means that
on a client system with a working KUBECONFIG pointing at your single
node cluster, you will be able to just download the kim binary
appropriate to your platform and run `kim build .` and it should Just
Work :TM:

New command hierarchy similar to `docker image build/push/pull/...`.
Just as with Docker, there are top level shortcuts such that `kim rmi`
is equivalent to calling `kim image rm`.

Ininstall, uninstall, and the as-of-yet-unimplemented info have been
moved under the `system` sub-command with no top-level shortcuts.
