+++ pkg/cli/command/image/pull/pull.go
@@ -1,0 +1,38 @@
+package pull
+
+import (
+	"github.com/pkg/errors"
+	"github.com/rancher/kim/pkg/client"
+	"github.com/rancher/kim/pkg/client/image"
+	wrangler "github.com/rancher/wrangler-cli"
+	"github.com/spf13/cobra"
+)
+
+const (
+	Use   = "pull [OPTIONS] IMAGE"
+	Short = "Pull an image"
+)
+
+func Command() *cobra.Command {
+	return wrangler.Command(&CommandSpec{}, cobra.Command{
+		Use:                   Use,
+		Short:                 Short,
+		DisableFlagsInUseLine: true,
+	})
+}
+
+type CommandSpec struct {
+	image.Pull
+}
+
+func (s *CommandSpec) Run(cmd *cobra.Command, args []string) error {
+	if len(args) != 1 {
+		return errors.New("exactly one argument is required")
+	}
+
+	k8s, err := client.DefaultConfig.Interface()
+	if err != nil {
+		return err
+	}
+	return s.Pull.Do(cmd.Context(), k8s, args[0])
+}
