+++ pkg/cli/command/image/push/push.go
@@ -1,0 +1,38 @@
+package push
+
+import (
+	"github.com/pkg/errors"
+	"github.com/rancher/kim/pkg/client"
+	"github.com/rancher/kim/pkg/client/image"
+	wrangler "github.com/rancher/wrangler-cli"
+	"github.com/spf13/cobra"
+)
+
+const (
+	Use   = "push [OPTIONS] IMAGE"
+	Short = "Push an image"
+)
+
+func Command() *cobra.Command {
+	return wrangler.Command(&CommandSpec{}, cobra.Command{
+		Use:                   Use,
+		Short:                 Short,
+		DisableFlagsInUseLine: true,
+	})
+}
+
+type CommandSpec struct {
+	image.Push
+}
+
+func (s *CommandSpec) Run(cmd *cobra.Command, args []string) error {
+	if len(args) != 1 {
+		return errors.New("exactly one argument is required")
+	}
+
+	k8s, err := client.DefaultConfig.Interface()
+	if err != nil {
+		return err
+	}
+	return s.Push.Do(cmd.Context(), k8s, args[0])
+}
