--- pkg/server/action/agent_other.go
@@ -1,9 +1,0 @@
-// +build !linux
-
-package action
-
-import "context"
-
-func (s *Agent) Run(ctx context.Context) error {
-	panic("not supported on this platform")
-}
